// frontend/src/app/dev/tables/cost-table/[...slug]/table/type.ts
export interface costTableRawDataProps {
    shopProductCardId: number; 
    candidate: boolean;
    shopProductName: string;
    shopProductImgUrl: string;
    productId: string;
    productUrl: string;
    brandName: string;
    originalPriceCurrency: string;
    originalPrice: number;
    taxReductionOriginalPrice: number;
    intlShipKorPrice: number;
    intlShipPrice: number;
    korPrice: number;
    usPrice: number;
    customFee: number;
    VATFee: number;
    totalPriceBeforeCardFee: number;
    totalPrice: number;
    sellPrice10P: number;
    sellPrice20P: number;
    cardFee: number;
    updatedAt: Date;
}